Refactoring for `ImmersedBoundaries` to take precedence over `Solvers` (#3847)
* try it out

* precompiles

* couple of bugfixes

* using inactive_node

* import regularize_boundary_condition

* flip the location

* flux form advection

* import conditional_flux

* disambiguation

* Update src/Coriolis/hydrostatic_spherical_coriolis.jl


* update importing

---------
