+++ src/Coriolis/hydrostatic_spherical_coriolis.jl
@@ -4,6 +4,7 @@ using Oceananigans.Advection: EnergyConserving, EnstrophyConserving
 using Oceananigans.BoundaryConditions
 using Oceananigans.Fields
 using Oceananigans.AbstractOperations: KernelFunctionOperation
+using Oceananigans.ImmersedBoundaries
 
 """
     struct ActiveCellEnstrophyConserving
@@ -49,6 +50,7 @@ Adapt.adapt_structure(to, coriolis::HydrostaticSphericalCoriolis) =
 
 @inline φᶠᶠᵃ(i, j, k, grid::LatitudeLongitudeGrid)        = φnode(j, grid, Face())
 @inline φᶠᶠᵃ(i, j, k, grid::OrthogonalSphericalShellGrid) = φnode(i, j, grid, Face(), Face())
+@inline φᶠᶠᵃ(i, j, k, grid::ImmersedBoundaryGrid)         = φᶠᶠᵃ(i, j, k, grid.underlying_grid)
 
 @inline fᶠᶠᵃ(i, j, k, grid, coriolis::HydrostaticSphericalCoriolis) =
     2 * coriolis.rotation_rate * hack_sind(φᶠᶠᵃ(i, j, k, grid))
