+++ src/ImmersedBoundaries/abstract_grid_fitted_boundary.jl
@@ -1,40 +1,6 @@
 abstract type AbstractGridFittedBoundary <: AbstractImmersedBoundary end
 
-import Oceananigans.TurbulenceClosures: ivd_upper_diagonal, ivd_lower_diagonal
-
 const GFIBG = ImmersedBoundaryGrid{<:Any, <:Any, <:Any, <:Any, <:Any, <:AbstractGridFittedBoundary}
-
-#####
-##### Implicit vertical diffusion
-#####
-##### For a center solver we have to check the interface "solidity" at faces k+1 in both the
-##### Upper diagonal and the Lower diagonal 
-##### (because of tridiagonal convention where lower_diagonal on row k is found at k-1)
-##### Same goes for the face solver, where we check at centers k in both Upper and lower diagonal
-#####
-
-@inline immersed_ivd_peripheral_node(i, j, k, ibg, ℓx, ℓy, ::Center) = immersed_peripheral_node(i, j, k+1, ibg, ℓx, ℓy, Face())
-@inline immersed_ivd_peripheral_node(i, j, k, ibg, ℓx, ℓy, ::Face)   = immersed_peripheral_node(i, j, k,   ibg, ℓx, ℓy, Center())
-
-# Extend the upper and lower diagonal functions of the batched tridiagonal solver
-
-for location in (:upper_, :lower_)
-    ordinary_func = Symbol(:ivd_ ,         location, :diagonal)
-    immersed_func = Symbol(:immersed_ivd_, location, :diagonal)
-    @eval begin
-        # Disambiguation
-        @inline $ordinary_func(i, j, k, ibg::IBG, closure, K, id, ℓx, ℓy, ℓz::Face, clock, Δt, κz) =
-                $immersed_func(i, j, k, ibg::IBG, closure, K, id, ℓx, ℓy, ℓz, clock, Δt, κz)
-
-        @inline $ordinary_func(i, j, k, ibg::IBG, closure, K, id, ℓx, ℓy, ℓz::Center, clock, Δt, κz) =
-                $immersed_func(i, j, k, ibg::IBG, closure, K, id, ℓx, ℓy, ℓz, clock, Δt, κz)
-
-        @inline $immersed_func(i, j, k, ibg::IBG, closure, K, id, ℓx, ℓy, ℓz, clock, Δt, κz) =
-            ifelse(immersed_ivd_peripheral_node(i, j, k, ibg, ℓx, ℓy, ℓz),
-                   zero(ibg),
-                   $ordinary_func(i, j, k, ibg.underlying_grid, closure, K, id, ℓx, ℓy, ℓz, clock, Δt, κz))
-    end
-end
 
 # Support for Flat grids
 # Note that instances of AbstractGridFittedBoundary should define _immersed_cell
